Add Mario stuck detection in the initial sequence collection
If mario was blocked somewhere, like a very high pipe,  a LOT of data be collected from that single place that will dominate the obs in training data. This explains the ghost pipe we saw in the reconstruction and may also casue the action policy learning fail. As policy trainer can't get reliable info from z_t and h_t, so the safest move is to blindly move right.