Add PercentileReturnNormalizer for robust return normalization with EMA
Till now, Mario can now learn to play and the recontructured game screen is clear. Two remaining issues: 1) In the reconstruction, background is clear, but Mario is blurry especially when Mario moves to the further right. 2) Mario's learning is not steady. Some obstacles he has learned to pass, but failed in the next training iteration